Move some test helpers to Workspaces.Test.Utilities
