+++ src/Workspaces/CoreTest/SolutionTests/ProjectSemanticVersionTests.cs
@@ -8,6 +8,7 @@ using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Xunit;
 using static Microsoft.CodeAnalysis.UnitTests.SolutionTestHelpers;
+using static Microsoft.CodeAnalysis.UnitTests.SolutionUtilities;
 
 namespace Microsoft.CodeAnalysis.UnitTests
 {
