+++ src/Workspaces/CoreTest/SolutionTests/SolutionTestHelpers.cs
@@ -32,18 +32,6 @@ namespace Microsoft.CodeAnalysis.UnitTests
             return workspace;
         }
 
-        public static Project AddEmptyProject(Solution solution, string languageName = LanguageNames.CSharp)
-        {
-            var id = ProjectId.CreateNewId();
-            return solution.AddProject(
-                ProjectInfo.Create(
-                    id,
-                    VersionStamp.Default,
-                    name: "TestProject",
-                    assemblyName: "TestProject",
-                    language: languageName)).GetRequiredProject(id);
-        }
-
 #nullable disable
 
         public static void TestProperty<T, TValue>(T instance, Func<T, TValue, T> factory, Func<T, TValue> getter, TValue validNonDefaultValue, bool defaultThrows = false)
